requirements.txt
streamlit
elasticsearch
python-dotenv
requests
gitpython